Added new code to files

+++ src/main/java/com/bank/retail_banking/entity/BankOffers.java
@@ -19,7 +19,7 @@ public class BankOffers {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer offerId;
+    private Long offerId;
 
     @Column(name = "offer_name",nullable = false,length = 100)
     private String offerName;
